requirements.txt
azure-ai-documentintelligence==1.0.0b1
azure-cosmos==4.5.1
azure-search-documents==11.4.0
openai==1.35.13
langchain-openai==0.1.15
python-dotenv
Flask
flask-cors
langchain